Revert "add builtin pipeline passes (#17985)"
This reverts commit aa3b9261e9bbdbd7dd2667cb18d166290691c7d4.
