--- editor/assets/default_renderpipeline/builtin-bloom-pass.ts.meta
@@ -1,9 +1,0 @@
-{
-  "ver": "4.0.24",
-  "importer": "typescript",
-  "imported": true,
-  "uuid": "5b9314e6-8955-435d-a14f-c8f2cdcdac44",
-  "files": [],
-  "subMetas": {},
-  "userData": {}
-}
