--- editor/assets/default_renderpipeline/builtin-fsr-pass.ts.meta
@@ -1,9 +1,0 @@
-{
-  "ver": "4.0.24",
-  "importer": "typescript",
-  "imported": true,
-  "uuid": "38163f91-9a44-4b7f-80cf-7876bb754bd4",
-  "files": [],
-  "subMetas": {},
-  "userData": {}
-}
