--- editor/assets/default_renderpipeline/builtin-fxaa-pass.ts.meta
@@ -1,9 +1,0 @@
-{
-  "ver": "4.0.24",
-  "importer": "typescript",
-  "imported": true,
-  "uuid": "bf9cf942-7180-4596-8da8-bfa3b95082c0",
-  "files": [],
-  "subMetas": {},
-  "userData": {}
-}
