+++ editor/assets/default_renderpipeline/builtin-pipeline.ts
@@ -172,7 +172,7 @@ function sortPipelinePassBuildersByRenderOrder(passBuilders: rendering.PipelineP
     });
 }
 
-export function addCopyToScreenPass(
+function addCopyToScreenPass(
     ppl: rendering.BasicPipeline,
     pplConfigs: Readonly<PipelineConfigs>,
     cameraConfigs: CameraConfigs,
@@ -979,6 +979,181 @@ export class BuiltinForwardPassBuilder implements rendering.PipelinePassBuilder 
     private readonly _reflectionProbeClearColor = new Vec3(0, 0, 0);
 }
 
+export interface BloomPassConfigs {
+    enableBloom: boolean;
+}
+
+export class BuiltinBloomPassBuilder implements rendering.PipelinePassBuilder {
+    getConfigOrder(): number {
+        return 0;
+    }
+    getRenderOrder(): number {
+        return 200;
+    }
+    configCamera(
+        camera: Readonly<renderer.scene.Camera>,
+        pipelineConfigs: Readonly<PipelineConfigs>,
+        cameraConfigs: CameraConfigs & BloomPassConfigs): void {
+        cameraConfigs.enableBloom
+            = cameraConfigs.settings.bloom.enabled
+            && !!cameraConfigs.settings.bloom.material;
+        if (cameraConfigs.enableBloom) {
+            ++cameraConfigs.remainingPasses;
+        }
+    }
+    windowResize(
+        ppl: rendering.BasicPipeline,
+        pplConfigs: Readonly<PipelineConfigs>,
+        cameraConfigs: CameraConfigs & BloomPassConfigs,
+        window: renderer.RenderWindow): void {
+        if (cameraConfigs.enableBloom) {
+            const id = window.renderWindowId;
+            let bloomWidth = cameraConfigs.width;
+            let bloomHeight = cameraConfigs.height;
+            for (let i = 0; i !== cameraConfigs.settings.bloom.iterations + 1; ++i) {
+                bloomWidth = Math.max(Math.floor(bloomWidth / 2), 1);
+                bloomHeight = Math.max(Math.floor(bloomHeight / 2), 1);
+                ppl.addRenderTarget(`BloomTex${id}_${i}`,
+                    cameraConfigs.radianceFormat, bloomWidth, bloomHeight);
+            }
+        }
+    }
+
+    setup(
+        ppl: rendering.BasicPipeline,
+        pplConfigs: Readonly<PipelineConfigs>,
+        cameraConfigs: CameraConfigs & BloomPassConfigs,
+        camera: renderer.scene.Camera,
+        context: PipelineContext,
+        prevRenderPass?: rendering.BasicRenderPassBuilder)
+        : rendering.BasicRenderPassBuilder | undefined {
+        if (!cameraConfigs.enableBloom) {
+            return prevRenderPass;
+        }
+
+        --cameraConfigs.remainingPasses;
+        assert(cameraConfigs.remainingPasses >= 0);
+        const id = camera.window.renderWindowId;
+        assert(!!cameraConfigs.settings.bloom.material);
+        return this._addKawaseDualFilterBloomPasses(
+            ppl, pplConfigs,
+            cameraConfigs,
+            cameraConfigs.settings,
+            cameraConfigs.settings.bloom.material,
+            id,
+            cameraConfigs.width,
+            cameraConfigs.height,
+            context.colorName);
+    }
+
+    private _addKawaseDualFilterBloomPasses(
+        ppl: rendering.BasicPipeline,
+        pplConfigs: Readonly<PipelineConfigs>,
+        cameraConfigs: CameraConfigs & Readonly<BloomPassConfigs>,
+        settings: PipelineSettings,
+        bloomMaterial: Material,
+        id: number,
+        width: number,
+        height: number,
+        radianceName: string,
+    ): rendering.BasicRenderPassBuilder {
+        const QueueHint = rendering.QueueHint;
+        // Based on Kawase Dual Filter Blur. Saves bandwidth on mobile devices.
+        // eslint-disable-next-line max-len
+        // https://community.arm.com/cfs-file/__key/communityserver-blogs-components-weblogfiles/00-00-00-20-66/siggraph2015_2D00_mmg_2D00_marius_2D00_slides.pdf
+
+        // Size: [prefilter(1/2), downsample(1/4), downsample(1/8), downsample(1/16), ...]
+        const iterations = settings.bloom.iterations;
+        const sizeCount = iterations + 1;
+        this._bloomWidths.length = sizeCount;
+        this._bloomHeights.length = sizeCount;
+        this._bloomWidths[0] = Math.max(Math.floor(width / 2), 1);
+        this._bloomHeights[0] = Math.max(Math.floor(height / 2), 1);
+        for (let i = 1; i !== sizeCount; ++i) {
+            this._bloomWidths[i] = Math.max(Math.floor(this._bloomWidths[i - 1] / 2), 1);
+            this._bloomHeights[i] = Math.max(Math.floor(this._bloomHeights[i - 1] / 2), 1);
+        }
+
+        // Bloom texture names
+        this._bloomTexNames.length = sizeCount;
+        for (let i = 0; i !== sizeCount; ++i) {
+            this._bloomTexNames[i] = `BloomTex${id}_${i}`;
+        }
+
+        // Setup bloom parameters
+        this._bloomParams.x = pplConfigs.useFloatOutput ? 1 : 0;
+        this._bloomParams.x = 0; // unused
+        this._bloomParams.z = settings.bloom.threshold;
+        this._bloomParams.w = settings.bloom.enableAlphaMask ? 1 : 0;
+
+        // Prefilter pass
+        const prefilterPass = ppl.addRenderPass(this._bloomWidths[0], this._bloomHeights[0], 'cc-bloom-prefilter');
+        prefilterPass.addRenderTarget(
+            this._bloomTexNames[0],
+            LoadOp.CLEAR,
+            StoreOp.STORE,
+            this._clearColorTransparentBlack,
+        );
+        prefilterPass.addTexture(radianceName, 'inputTexture');
+        prefilterPass.setVec4('g_platform', pplConfigs.platform);
+        prefilterPass.setVec4('bloomParams', this._bloomParams);
+        prefilterPass
+            .addQueue(QueueHint.OPAQUE)
+            .addFullscreenQuad(bloomMaterial, 0);
+
+        // Downsample passes
+        for (let i = 1; i !== sizeCount; ++i) {
+            const downPass = ppl.addRenderPass(this._bloomWidths[i], this._bloomHeights[i], 'cc-bloom-downsample');
+            downPass.addRenderTarget(this._bloomTexNames[i], LoadOp.CLEAR, StoreOp.STORE, this._clearColorTransparentBlack);
+            downPass.addTexture(this._bloomTexNames[i - 1], 'bloomTexture');
+            this._bloomTexSize.x = this._bloomWidths[i - 1];
+            this._bloomTexSize.y = this._bloomHeights[i - 1];
+            downPass.setVec4('g_platform', pplConfigs.platform);
+            downPass.setVec4('bloomTexSize', this._bloomTexSize);
+            downPass
+                .addQueue(QueueHint.OPAQUE)
+                .addFullscreenQuad(bloomMaterial, 1);
+        }
+
+        // Upsample passes
+        for (let i = iterations; i-- > 0;) {
+            const upPass = ppl.addRenderPass(this._bloomWidths[i], this._bloomHeights[i], 'cc-bloom-upsample');
+            upPass.addRenderTarget(this._bloomTexNames[i], LoadOp.CLEAR, StoreOp.STORE, this._clearColorTransparentBlack);
+            upPass.addTexture(this._bloomTexNames[i + 1], 'bloomTexture');
+            this._bloomTexSize.x = this._bloomWidths[i + 1];
+            this._bloomTexSize.y = this._bloomHeights[i + 1];
+            upPass.setVec4('g_platform', pplConfigs.platform);
+            upPass.setVec4('bloomTexSize', this._bloomTexSize);
+            upPass
+                .addQueue(QueueHint.OPAQUE)
+                .addFullscreenQuad(bloomMaterial, 2);
+        }
+
+        // Combine pass
+        const combinePass = ppl.addRenderPass(width, height, 'cc-bloom-combine');
+        combinePass.addRenderTarget(radianceName, LoadOp.LOAD, StoreOp.STORE);
+        combinePass.addTexture(this._bloomTexNames[0], 'bloomTexture');
+        combinePass.setVec4('g_platform', pplConfigs.platform);
+        combinePass.setVec4('bloomParams', this._bloomParams);
+        combinePass
+            .addQueue(QueueHint.BLEND)
+            .addFullscreenQuad(bloomMaterial, 3);
+
+        if (cameraConfigs.remainingPasses === 0) {
+            return addCopyToScreenPass(ppl, pplConfigs, cameraConfigs, radianceName);
+        } else {
+            return combinePass;
+        }
+    }
+    // Bloom
+    private readonly _clearColorTransparentBlack = new Color(0, 0, 0, 0);
+    private readonly _bloomParams = new Vec4(0, 0, 0, 0);
+    private readonly _bloomTexSize = new Vec4(0, 0, 0, 0);
+    private readonly _bloomWidths: Array<number> = [];
+    private readonly _bloomHeights: Array<number> = [];
+    private readonly _bloomTexNames: Array<string> = [];
+}
+
 export interface ToneMappingPassConfigs {
     enableToneMapping: boolean;
     enableColorGrading: boolean;
@@ -1037,7 +1212,7 @@ export class BuiltinToneMappingPassBuilder implements rendering.PipelinePassBuil
         assert(cameraConfigs.remainingPasses >= 0);
         if (cameraConfigs.remainingPasses === 0) {
             return this._addCopyAndTonemapPass(ppl, pplConfigs, cameraConfigs,
-                cameraConfigs.nativeWidth, cameraConfigs.nativeHeight,
+                cameraConfigs.width, cameraConfigs.height,
                 context.colorName, cameraConfigs.colorName);
         } else {
             const id = cameraConfigs.renderWindowId;
@@ -1105,6 +1280,214 @@ export class BuiltinToneMappingPassBuilder implements rendering.PipelinePassBuil
     private readonly _colorGradingTexSize = new Vec2(0, 0);
 }
 
+export interface FXAAPassConfigs {
+    enableFXAA: boolean;
+}
+
+export class BuiltinFXAAPassBuilder implements rendering.PipelinePassBuilder {
+    getConfigOrder(): number {
+        return 0;
+    }
+    getRenderOrder(): number {
+        return 400;
+    }
+    configCamera(
+        camera: Readonly<renderer.scene.Camera>,
+        pplConfigs: Readonly<PipelineConfigs>,
+        cameraConfigs: CameraConfigs & FXAAPassConfigs): void {
+        cameraConfigs.enableFXAA
+            = cameraConfigs.settings.fxaa.enabled
+            && !!cameraConfigs.settings.fxaa.material;
+        if (cameraConfigs.enableFXAA) {
+            ++cameraConfigs.remainingPasses;
+        }
+    }
+    setup(
+        ppl: rendering.BasicPipeline,
+        pplConfigs: Readonly<PipelineConfigs>,
+        cameraConfigs: CameraConfigs & FXAAPassConfigs,
+        camera: renderer.scene.Camera,
+        context: PipelineContext,
+        prevRenderPass?: rendering.BasicRenderPassBuilder)
+        : rendering.BasicRenderPassBuilder | undefined {
+        if (!cameraConfigs.enableFXAA) {
+            return prevRenderPass;
+        }
+        --cameraConfigs.remainingPasses;
+        assert(cameraConfigs.remainingPasses >= 0);
+
+        const id = cameraConfigs.renderWindowId;
+        const ldrColorPrefix = cameraConfigs.enableShadingScale
+            ? `ScaledLdrColor`
+            : `LdrColor`;
+        const ldrColorName = getPingPongRenderTarget(context.colorName, ldrColorPrefix, id);
+
+        assert(!!cameraConfigs.settings.fxaa.material);
+        if (cameraConfigs.remainingPasses === 0) {
+            if (cameraConfigs.enableShadingScale) {
+                this._addFxaaPass(ppl, pplConfigs,
+                    cameraConfigs.settings.fxaa.material,
+                    cameraConfigs.width,
+                    cameraConfigs.height,
+                    context.colorName,
+                    ldrColorName);
+                return addCopyToScreenPass(ppl, pplConfigs, cameraConfigs, ldrColorName);
+            } else {
+                assert(cameraConfigs.width === cameraConfigs.nativeWidth);
+                assert(cameraConfigs.height === cameraConfigs.nativeHeight);
+                return this._addFxaaPass(ppl, pplConfigs,
+                    cameraConfigs.settings.fxaa.material,
+                    cameraConfigs.width,
+                    cameraConfigs.height,
+                    context.colorName,
+                    cameraConfigs.colorName);
+            }
+        } else {
+            const inputColorName = context.colorName;
+            context.colorName = ldrColorName;
+            const lastPass = this._addFxaaPass(ppl, pplConfigs,
+                cameraConfigs.settings.fxaa.material,
+                cameraConfigs.width,
+                cameraConfigs.height,
+                inputColorName,
+                ldrColorName);
+            return lastPass;
+        }
+    }
+    private _addFxaaPass(
+        ppl: rendering.BasicPipeline,
+        pplConfigs: Readonly<PipelineConfigs>,
+        fxaaMaterial: Material,
+        width: number,
+        height: number,
+        ldrColorName: string,
+        colorName: string,
+    ): rendering.BasicRenderPassBuilder {
+        this._fxaaParams.x = width;
+        this._fxaaParams.y = height;
+        this._fxaaParams.z = 1 / width;
+        this._fxaaParams.w = 1 / height;
+
+        const pass = ppl.addRenderPass(width, height, 'cc-fxaa');
+        pass.addRenderTarget(colorName, LoadOp.CLEAR, StoreOp.STORE, sClearColorTransparentBlack);
+        pass.addTexture(ldrColorName, 'sceneColorMap');
+        pass.setVec4('g_platform', pplConfigs.platform);
+        pass.setVec4('texSize', this._fxaaParams);
+        pass.addQueue(rendering.QueueHint.OPAQUE)
+            .addFullscreenQuad(fxaaMaterial, 0);
+        return pass;
+    }
+    // FXAA
+    private readonly _fxaaParams = new Vec4(0, 0, 0, 0);
+}
+
+export interface FSRPassConfigs {
+    enableFSR: boolean;
+}
+
+export class BuiltinFsrPassBuilder implements rendering.PipelinePassBuilder {
+    getConfigOrder(): number {
+        return 0;
+    }
+    getRenderOrder(): number {
+        return 500;
+    }
+    configCamera(
+        camera: Readonly<renderer.scene.Camera>,
+        pplConfigs: Readonly<PipelineConfigs>,
+        cameraConfigs: CameraConfigs & FSRPassConfigs): void {
+        // FSR (Depend on Shading scale)
+        cameraConfigs.enableFSR = cameraConfigs.settings.fsr.enabled
+            && !!cameraConfigs.settings.fsr.material
+            && cameraConfigs.enableShadingScale
+            && cameraConfigs.shadingScale < 1.0;
+
+        if (cameraConfigs.enableFSR) {
+            ++cameraConfigs.remainingPasses;
+        }
+    }
+    setup(
+        ppl: rendering.BasicPipeline,
+        pplConfigs: Readonly<PipelineConfigs>,
+        cameraConfigs: CameraConfigs & FSRPassConfigs,
+        camera: renderer.scene.Camera,
+        context: PipelineContext,
+        prevRenderPass?: rendering.BasicRenderPassBuilder)
+        : rendering.BasicRenderPassBuilder | undefined {
+        if (!cameraConfigs.enableFSR) {
+            return prevRenderPass;
+        }
+        --cameraConfigs.remainingPasses;
+
+        const inputColorName = context.colorName;
+        const outputColorName
+            = cameraConfigs.remainingPasses === 0
+                ? cameraConfigs.colorName
+                : getPingPongRenderTarget(context.colorName, 'UiColor', cameraConfigs.renderWindowId);
+        context.colorName = outputColorName;
+
+        assert(!!cameraConfigs.settings.fsr.material);
+        return this._addFsrPass(ppl, pplConfigs, cameraConfigs,
+            cameraConfigs.settings,
+            cameraConfigs.settings.fsr.material,
+            cameraConfigs.renderWindowId,
+            cameraConfigs.width,
+            cameraConfigs.height,
+            inputColorName,
+            cameraConfigs.nativeWidth,
+            cameraConfigs.nativeHeight,
+            outputColorName);
+    }
+    private _addFsrPass(
+        ppl: rendering.BasicPipeline,
+        pplConfigs: Readonly<PipelineConfigs>,
+        cameraConfigs: CameraConfigs & FSRPassConfigs,
+        settings: PipelineSettings,
+        fsrMaterial: Material,
+        id: number,
+        width: number,
+        height: number,
+        inputColorName: string,
+        nativeWidth: number,
+        nativeHeight: number,
+        outputColorName: string,
+    ): rendering.BasicRenderPassBuilder {
+        this._fsrTexSize.x = width;
+        this._fsrTexSize.y = height;
+        this._fsrTexSize.z = nativeWidth;
+        this._fsrTexSize.w = nativeHeight;
+        this._fsrParams.x = clamp(1.0 - settings.fsr.sharpness, 0.02, 0.98);
+
+        const uiColorPrefix = 'UiColor';
+
+        const fsrColorName = getPingPongRenderTarget(outputColorName, uiColorPrefix, id);
+
+        const easuPass = ppl.addRenderPass(nativeWidth, nativeHeight, 'cc-fsr-easu');
+        easuPass.addRenderTarget(fsrColorName, LoadOp.CLEAR, StoreOp.STORE, sClearColorTransparentBlack);
+        easuPass.addTexture(inputColorName, 'outputResultMap');
+        easuPass.setVec4('g_platform', pplConfigs.platform);
+        easuPass.setVec4('fsrTexSize', this._fsrTexSize);
+        easuPass
+            .addQueue(rendering.QueueHint.OPAQUE)
+            .addFullscreenQuad(fsrMaterial, 0);
+
+        const rcasPass = ppl.addRenderPass(nativeWidth, nativeHeight, 'cc-fsr-rcas');
+        rcasPass.addRenderTarget(outputColorName, LoadOp.CLEAR, StoreOp.STORE, sClearColorTransparentBlack);
+        rcasPass.addTexture(fsrColorName, 'outputResultMap');
+        rcasPass.setVec4('g_platform', pplConfigs.platform);
+        rcasPass.setVec4('fsrTexSize', this._fsrTexSize);
+        rcasPass.setVec4('fsrParams', this._fsrParams);
+        rcasPass
+            .addQueue(rendering.QueueHint.OPAQUE)
+            .addFullscreenQuad(fsrMaterial, 1);
+
+        return rcasPass;
+    }
+    // FSR
+    private readonly _fsrParams = new Vec4(0, 0, 0, 0);
+    private readonly _fsrTexSize = new Vec4(0, 0, 0, 0);
+}
+
 export class BuiltinUiPassBuilder implements rendering.PipelinePassBuilder {
     getConfigOrder(): number {
         return 0;
@@ -1115,7 +1498,7 @@ export class BuiltinUiPassBuilder implements rendering.PipelinePassBuilder {
     setup(
         ppl: rendering.BasicPipeline,
         pplConfigs: Readonly<PipelineConfigs>,
-        cameraConfigs: CameraConfigs,
+        cameraConfigs: CameraConfigs & FSRPassConfigs,
         camera: renderer.scene.Camera,
         context: PipelineContext,
         prevRenderPass?: rendering.BasicRenderPassBuilder)
@@ -1142,7 +1525,10 @@ if (rendering) {
     class BuiltinPipelineBuilder implements rendering.PipelineBuilder {
         private readonly _pipelineEvent: PipelineEventProcessor = cclegacy.director.root.pipelineEvent as PipelineEventProcessor;
         private readonly _forwardPass = new BuiltinForwardPassBuilder();
+        private readonly _bloomPass = new BuiltinBloomPassBuilder();
         private readonly _toneMappingPass = new BuiltinToneMappingPassBuilder();
+        private readonly _fxaaPass = new BuiltinFXAAPassBuilder();
+        private readonly _fsrPass = new BuiltinFsrPassBuilder();
         private readonly _uiPass = new BuiltinUiPassBuilder();
         // Internal cached resources
         private readonly _clearColor = new Color(0, 0, 0, 1);
@@ -1192,7 +1578,20 @@ if (rendering) {
             }
 
             passBuilders.push(this._forwardPass);
+
+            if (settings.bloom.enabled) {
+                passBuilders.push(this._bloomPass);
+            }
+
             passBuilders.push(this._toneMappingPass);
+
+            if (settings.fxaa.enabled) {
+                passBuilders.push(this._fxaaPass);
+            }
+
+            if (settings.fsr.enabled) {
+                passBuilders.push(this._fsrPass);
+            }
             passBuilders.push(this._uiPass);
         }
